requirements.txt
# Django y REST framework
Django>=5.0.0
djangorestframework>=3.14.0

# Documentación API
drf-spectacular>=0.27.0

# Base de datos
mysqlclient>=2.2.0

# Autenticación y seguridad
PyJWT>=2.8.0
bcrypt>=4.1.0
django-cors-headers>=4.3.0

# Utilidades
python-dotenv>=1.0.0

# Formateo y linting (desarrollo)
black>=24.0.0
flake8>=7.0.0 
python-decouple 
boto3 
django-storages
Pillow #compress images
pip install drf-extra-fields
